Allow custom configurations using the built-in settings.
The script automatically looks for a config file, but inside
Sublime Text 2, it wasn't pulling from the right place. This just
changes it to look in the same folder as the script (inside the plugin
folder).

This should only be used for the non-WP `phptidy.php` script, because
the plugin already handles the custom WordPress stylings.